ELMMRUI
We evaluate the effects of unemployment insurance variations in a general equilibriumoccupational choice model of entrepreneurship.  We establish that the occupational flowfrom  unemployment  to  entrepreneurship  is  remarkably  sensitive  to  the  unemploymentinsurance generosity, corroborating our empirical findings.  Beyond direct effects on un-employment, we find large reallocations between employment and entrepreneurship rel-ative  to  changes  in  the  generosity.   They  contribute  to  an  empirically  consistent  stableaggregate employment rate, despite the increasing unemployment.  We show that anin-surance coverage effect,  i.e.   a  change  in  the  relative  riskiness  between  occupations  withrespect to the generosity, is a key driver of the results.